+++ packages/core/src/components/widgets/SelectWidget.tsx
@@ -40,7 +40,7 @@ function SelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F extend
   placeholder,
   htmlName,
 }: WidgetProps<T, S, F>) {
-  const { enumOptions, enumDisabled, emptyValue: optEmptyVal } = options;
+  const { enumOptions, enumDisabled, emptyValue: optEmptyVal, optgroups } = options;
   const emptyValue = multiple ? [] : '';
 
   const handleFocus = useCallback(
@@ -68,7 +68,37 @@ function SelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F extend
   );
 
   const selectedIndexes = enumOptionsIndexForValue<S>(value, enumOptions, multiple);
-  const showPlaceholderOption = !multiple && schema.default === undefined;
+
+  let opts = null
+  if (optgroups) {
+    let enumOptionFromValue = new Map()
+    enumOptions.forEach((enumOption, i) => {
+      enumOptionFromValue.set(enumOption.value, [enumOption, i])
+    })
+    opts = Object.keys(optgroups).map(optgroup => {
+      return (
+        <optgroup key={optgroup} label={optgroup}>{optgroups[optgroup].map(value => {
+          if (!enumOptionFromValue.has(value)) return null
+          const disabled = enumDisabled && enumDisabled.indexOf(value) !== -1;
+          let [{ label }, i] = enumOptionFromValue.get(value)
+          return (
+            <option key={i} value={String(i)} disabled={disabled}>
+              {label}
+            </option>
+          );
+        })}</optgroup>
+      )
+    })
+  } else {
+    opts = Array.isArray(enumOptions) && enumOptions.map(({ value, label }, i) => {
+      const disabled = enumDisabled && enumDisabled.indexOf(value) !== -1;
+      return (
+        <option key={i} value={String(i)} disabled={disabled}>
+          {label}
+        </option>
+      );
+    });
+  }
 
   return (
     <select
@@ -86,16 +116,8 @@ function SelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F extend
       onChange={handleChange}
       aria-describedby={ariaDescribedByIds(id)}
     >
-      {showPlaceholderOption && <option value=''>{placeholder}</option>}
-      {Array.isArray(enumOptions) &&
-        enumOptions.map(({ value, label }, i) => {
-          const disabled = enumDisabled && enumDisabled.indexOf(value) !== -1;
-          return (
-            <option key={i} value={String(i)} disabled={disabled}>
-              {label}
-            </option>
-          );
-        })}
+      {!multiple && schema.default === undefined && <option value=''>{placeholder}</option>}
+      {opts}
     </select>
   );
 }