+++ packages/playground/src/samples/widgets.tsx
@@ -168,6 +168,11 @@ const widgets: Sample = {
           },
         ],
       },
+      selectWidgetOptions3: {
+        title: 'Custom select widget with options grouped by optgroups',
+        type: 'string',
+        enum: ['foo', 'bar'],
+      },
     },
   },
   uiSchema: {
@@ -284,6 +289,14 @@ const widgets: Sample = {
         backgroundColor: 'pink',
       },
     },
+    selectWidgetOptions3: {
+      'ui:options': {
+        'optgroups': {
+          'lipsum': ['foo'],
+          'dolorem': ['bar']
+        }
+      }
+    }
   },
   formData: {
     stringFormats: {